interview/Demo/tst/com/peter/left/interview/dataStructure/problems/LongestSubstringWithoutRepeatingCharacters/SolutionTest.java
package com.peter.left.interview.dataStructure.problems.LongestSubstringWithoutRepeatingCharacters;

import static org.hamcrest.Matchers.equalTo;
import static org.junit.Assert.assertThat;

import java.util.Collections;

import org.junit.Before;
import org.junit.Test;



public class SolutionTest
{
    private Solution solution;
    
    @Before
    public void setUp()
    {
        solution = new Solution();
    }
    
    @Test
    public void test1()
    {
        final int[] values = {1};
        assertThat(solution.lengthOfLongestSubstring("wlrbbmqbhcdarzowkkyhiddqscdxrjmowfrxsjybldbefsarcbynecdyggxxpklorellnmpapqfwkhopkmco"), 
                equalTo(12));
    }
}
